move discretize.py to factors.continuous
