--- pgmpy/discretize/__init__.py
@@ -1,5 +1,0 @@
-from .discretize import BaseDiscretizer, RoundingDiscretizer, UnbiasedDiscretizer
-
-__all__ = ['BaseDiscretizer',
-           'RoundingDiscretizer',
-           'UnbiasedDiscretizer']
