+++ pgmpy/factors/continuous/discretize.py
