+++ pgmpy/tests/test_factors/test_continuous/test_ContinuousNode.py
@@ -5,9 +5,9 @@ import numpy as np
 from scipy.stats import norm, t, gamma
 
 from pgmpy.factors.continuous import ContinuousNode
-from pgmpy.discretize import BaseDiscretizer
-from pgmpy.discretize import RoundingDiscretizer
-from pgmpy.discretize import UnbiasedDiscretizer
+from pgmpy.factors.continuous import BaseDiscretizer
+from pgmpy.factors.continuous import RoundingDiscretizer
+from pgmpy.factors.continuous import UnbiasedDiscretizer
 
 
 class TestContinuousNodeInit(unittest.TestCase):
